+++ CHANGELOG.md
@@ -3,9 +3,9 @@
 ## 4.0 GUI and Plugin Refactoring
 
 - 2023-03-07 4.0.48
-    - don't permanently interpolate clipboard measurements during export
+    - temporarily interpolate clipboard measurements during export
     - display "sensor stabilizing" message on SiG
-    - Controller sends wasatch.StatusMessage with marquee_info or marquee_error to Marquee, not Toast
+    - display wasatch.StatusMessage via Marquee, not Toast
 - 2023-03-07 4.0.47
     - fixed export bug
 - 2023-03-01 4.0.46